backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend of Perdus"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "firebase-admin>=7.1.0",
    "google-genai>=1.60.0",
    "numpy==1.26.4",
    "pillow>=12.1.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.1.2",
    "torch==2.2.0",
    "uvicorn>=0.40.0",
]
